chore(suite-sync): rename createServices
